+++ IsekaiMod/Content/Features/IsekaiProtagonist/CharacterDevelopment/EnergyImmunitySelection.cs
@@ -121,6 +121,7 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist.CharacterDevelopment
                 bp.m_Features = EnergyImmunityList;
                 bp.m_AllFeatures = EnergyImmunityList;
             });
+            CharacterDevelopmentSelection.AddToSelection(EnergyImmunitySelection);
         }
     }
 }