+++ IsekaiMod/Content/Features/IsekaiProtagonist/CharacterDevelopment/GammaStrike.cs
@@ -4,14 +4,15 @@ using Kingmaker.Blueprints.Classes;
 using Kingmaker.Designers.Mechanics.Facts;
 using Kingmaker.Enums.Damage;
 using Kingmaker.UnitLogic.FactLogic;
+using UnityEngine;
 
 namespace IsekaiMod.Content.Features.IsekaiProtagonist.CharacterDevelopment
 {
     class GammaStrike
     {
+        private static readonly Sprite Icon_BladeSense = Resources.GetBlueprint<BlueprintFeature>("112bf4c6943097942b24eadfa750215f").m_Icon;
         public static void Add()
         {
-            var Icon_BladeSense = Resources.GetBlueprint<BlueprintFeature>("112bf4c6943097942b24eadfa750215f").m_Icon;
             var GammaStrike = Helpers.CreateBlueprint<BlueprintFeature>("GammaStrike", bp => {
                 bp.SetName("Gamma Strike");
                 bp.SetDescription("Your attacks ignore concealment and are treated as adamantite for the purpose of overcoming {g|Encyclopedia:Damage_Reduction}damage reduction{/g}.");
@@ -24,6 +25,8 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist.CharacterDevelopment
                 bp.Ranks = 1;
                 bp.IsClassFeature = true;
             });
+
+            CharacterDevelopmentSelection.AddToSelection(GammaStrike);
         }
     }
 }