+++ IsekaiMod/Content/Features/IsekaiProtagonist/CharacterDevelopment/MundaneAura.cs
@@ -2,14 +2,15 @@
 using IsekaiMod.Utilities;
 using Kingmaker.Blueprints.Classes;
 using Kingmaker.UnitLogic.FactLogic;
+using UnityEngine;
 
 namespace IsekaiMod.Content.Features.IsekaiProtagonist.CharacterDevelopment
 {
     class MundaneAura
     {
+        private static readonly Sprite Icon_BardLoreMaster = Resources.GetBlueprint<BlueprintFeature>("4bea694e79a87cd4d8c14fb91578059e").m_Icon;
         public static void Add()
         {
-            var Icon_BardLoreMaster = Resources.GetBlueprint<BlueprintFeature>("4bea694e79a87cd4d8c14fb91578059e").m_Icon;
             var MundaneAura = Helpers.CreateBlueprint<BlueprintFeature>("MundaneAura", bp => {
                 bp.SetName("Mundane Aura");
                 bp.SetDescription("You emit a small aura of mundanity, giving you immunity to Sneak attack damage and {g|Encyclopedia:Critical}critical hits{/g}.");
@@ -19,6 +20,8 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist.CharacterDevelopment
                 bp.Ranks = 1;
                 bp.IsClassFeature = true;
             });
+
+            CharacterDevelopmentSelection.AddToSelection(MundaneAura);
         }
     }
 }