+++ IsekaiMod/Content/Features/IsekaiProtagonist/CharacterDevelopment/TrainingMontage.cs
@@ -1,55 +1,62 @@
 ﻿using IsekaiMod.Extensions;
 using IsekaiMod.Utilities;
 using Kingmaker.Blueprints.Classes;
+using Kingmaker.Blueprints.Classes.Prerequisites;
 using Kingmaker.EntitySystem.Stats;
 using Kingmaker.Enums;
 using Kingmaker.UnitLogic.Abilities.Blueprints;
 using Kingmaker.UnitLogic.FactLogic;
+using UnityEngine;
 
-namespace IsekaiMod.Content.Features.IsekaiProtagonist.TrainingArc
+namespace IsekaiMod.Content.Features.IsekaiProtagonist.CharacterDevelopment
 {
     class TrainingMontage
     {
+        private static readonly Sprite Icon_LegendaryProportions = Resources.GetBlueprint<BlueprintAbility>("da1b292d91ba37948893cdbe9ea89e28").m_Icon;
         public static void Add()
         {
-            var Icon_LegendaryProportions = Resources.GetBlueprint<BlueprintAbility>("da1b292d91ba37948893cdbe9ea89e28").m_Icon;
             var TrainingMontage = Helpers.CreateBlueprint<BlueprintFeature>("TrainingMontage", bp => {
                 bp.SetName("Training Montage");
-                bp.SetDescription("After extensive training sessions, you gain a +8 insight bonus to all attributes.");
+                bp.SetDescription("After extensive training, you gain a +10 bonus to all attributes.");
                 bp.AddComponent<AddStatBonus>(c => {
-                    c.Descriptor = ModifierDescriptor.Insight;
+                    c.Descriptor = ModifierDescriptor.UntypedStackable;
                     c.Stat = StatType.Strength;
-                    c.Value = 8;
+                    c.Value = 10;
                 });
                 bp.AddComponent<AddStatBonus>(c => {
-                    c.Descriptor = ModifierDescriptor.Insight;
+                    c.Descriptor = ModifierDescriptor.UntypedStackable;
                     c.Stat = StatType.Dexterity;
-                    c.Value = 8;
+                    c.Value = 10;
                 });
                 bp.AddComponent<AddStatBonus>(c => {
-                    c.Descriptor = ModifierDescriptor.Insight;
+                    c.Descriptor = ModifierDescriptor.UntypedStackable;
                     c.Stat = StatType.Constitution;
-                    c.Value = 8;
+                    c.Value = 10;
                 });
                 bp.AddComponent<AddStatBonus>(c => {
-                    c.Descriptor = ModifierDescriptor.Insight;
+                    c.Descriptor = ModifierDescriptor.UntypedStackable;
                     c.Stat = StatType.Intelligence;
-                    c.Value = 8;
+                    c.Value = 10;
                 });
                 bp.AddComponent<AddStatBonus>(c => {
-                    c.Descriptor = ModifierDescriptor.Insight;
+                    c.Descriptor = ModifierDescriptor.UntypedStackable;
                     c.Stat = StatType.Wisdom;
-                    c.Value = 8;
+                    c.Value = 10;
                 });
                 bp.AddComponent<AddStatBonus>(c => {
-                    c.Descriptor = ModifierDescriptor.Insight;
+                    c.Descriptor = ModifierDescriptor.UntypedStackable;
                     c.Stat = StatType.Charisma;
-                    c.Value = 8;
+                    c.Value = 10;
+                });
+                bp.AddComponent<PrerequisiteCharacterLevel>(c => {
+                    c.Level = 10;
                 });
                 bp.m_Icon = Icon_LegendaryProportions;
                 bp.Ranks = 1;
                 bp.IsClassFeature = true;
             });
+
+            CharacterDevelopmentSelection.AddToSelection(TrainingMontage);
         }
     }
 }