+++ IsekaiMod/Content/Features/IsekaiProtagonist/Villain/CorruptAuraFeature.cs
@@ -101,7 +101,7 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist.Villain
             });
             var CorruptAuraFeature = Helpers.CreateBlueprint<BlueprintFeature>("CorruptAuraFeature", bp => {
                 bp.SetName("Corrupt Aura");
-                bp.SetDescription("At 10th level, allies within 40 feet of the Villain has a +4 profane bonus to attack, damage, AC and saving throws. "
+                bp.SetDescription("At 9th level, allies within 40 feet of the Villain has a +4 profane bonus to attack, damage, AC and saving throws. "
                     + "Their attacks are treated as evil for the purpose of overcoming {g|Encyclopedia:Damage_Reduction}damage reduction{/g}.");
                 bp.m_Icon = Icon_Corrupt_Aura;
                 bp.Ranks = 1;